feat(express): remove unnecessary default export
